requirements.txt
pycryptodome
questionary
gitpython
python-gnupg
llama-cpp-python
